build(DESCRIPTION): cleanup and add catboost
* fix: add catboost to DESCRIPTION

Because of a bug in pak, the package installation in the CI
failed when catboost was in the DESCRIPTION, because of a bug
in pak. With pak 4.0.0 that should be resolved.

https://github.com/r-lib/pak/issues/385

* chore(tests): remove unneeded setup and teardown files

Initially the idea was to enable additional check for partial
argument matching. But this throws too many warnings as the
dependencies violate this.

* build(remotes): Update remotes for aorsf and catboost

* Fix catboost remote to latest release
* Use the CRAN version of aorsf again

* ci(cron): Run rcmdcheck every week

This is useful because it prepares the cache.