feat(learner): Priority Lasso (#266)
* priority lasso regr, classif and surv learners added

* fix: add catboost to DESCRIPTION

Because of a bug in pak, the package installation in the CI
failed when catboost was in the DESCRIPTION, because of a bug
in pak. With pak 4.0.0 that should be resolved.

https://github.com/r-lib/pak/issues/385

* chore(tests): remove unneeded setup and teardown files

Initially the idea was to enable additional check for partial
argument matching. But this throws too many warnings as the
dependencies violate this.

* build(remotes): Update remotes for aorsf and catboost

* Fix catboost remote to latest release
* Use the CRAN version of aorsf again

* ci(cron): Run rcmdcheck every week

This is useful because it prepares the cache.

* fix(vignette): fix error

* fix(priority_lasso): fix learner bugs and improves tests

The previous tests from the author were to weak and did not
catch the errors in the implementations.
With a workaround to provide task-dependent parameters
(quite hacky), we could still run the autotests and correct
the errors in the implementation.

* style(prioritlasso): address linting issues

* fix(syntax): solve merge conflict

* tests(priority_lasso): add missing parameter

* feat(priority_lasso): add prob response type to classif

* chore(tests): correct comments in parameter tests

* fix(tests): minor correction

* fix(priority_lasso): fix test

Co-authored-by: HarutyunyanLiana <lh.lianaharutyunyan@gmail.com>